requirements.txt
streamlit
pandas
plotly
Pillow
opencv-python
numpy
pytesseract
SpeechRecognition
textblob